requirements.txt
streamlit
openai
numpy
requests
python-dotenv
